feat(container-fs): implement withOutputChmod for host-writable output and add tests
