+++ test/s57-converter.test.ts
@@ -18,7 +18,6 @@ const {
   bandClampedMaxzoom,
   buildLayerManifest,
   buildTippecanoeCommand,
-  withOutputChmod,
   TIPPECANOE_LAYER_MANIFEST,
   surfaceExportErrorsIfEmpty,
   wantedMbtilesName,
@@ -537,41 +536,6 @@ describe('buildTippecanoeCommand (argv stays small regardless of layer count)', 
     assert.ok(
       typeof TIPPECANOE_LAYER_MANIFEST === 'string' && TIPPECANOE_LAYER_MANIFEST.length > 0
     );
-  });
-});
-
-// gdaladdo/tile-join (unlike tippecanoe) run as plain argv today; this is
-// what wraps them in the same bash -c + trailing-chmod shape so their
-// output is host-writable afterward too — see buildTippecanoeCommand above
-// for why that matters.
-describe('withOutputChmod (gdaladdo/tile-join output becomes host-writable)', () => {
-  it('returns a 3-element bash -c command running the original argv then a chmod', () => {
-    const cmd = withOutputChmod(
-      ['gdaladdo', '-r', 'average', '/output/x.mbtiles'],
-      '/output/x.mbtiles'
-    );
-    assert.strictEqual(cmd.length, 3);
-    assert.strictEqual(cmd[0], 'bash');
-    assert.strictEqual(cmd[1], '-c');
-    const lines = cmd[2].split('\n');
-    const argvLine = lines.findIndex((l: string) => l.includes('gdaladdo'));
-    const chmodLine = lines.findIndex((l: string) => l.includes('chmod 666'));
-    assert.ok(argvLine >= 0 && chmodLine >= 0);
-    assert.ok(chmodLine > argvLine, 'chmod must run after the original command, not before');
-  });
-
-  it('shell-quotes both the argv and the chmod target so paths with spaces survive', () => {
-    const cmd = withOutputChmod(
-      ['gdaladdo', '-r', 'average', '/out put/x.mbtiles'],
-      '/out put/x.mbtiles'
-    );
-    assert.match(cmd[2], /'\/out put\/x\.mbtiles'/);
-    assert.match(cmd[2], /chmod 666 '\/out put\/x\.mbtiles'/);
-  });
-
-  it('runs `set -e` first so a failing original command skips the chmod', () => {
-    const cmd = withOutputChmod(['gdaladdo', '/output/x.mbtiles'], '/output/x.mbtiles');
-    assert.strictEqual(cmd[2].split('\n')[0], 'set -e');
   });
 });
